backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
sqlalchemy==2.0.46
alembic==1.18.4
pyjwt==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.22
langgraph==1.0.9
langchain==1.2.10
langchain-openai==1.1.10
numpy==1.26.4
scipy==1.10.1
scikit-learn==1.2.2
joblib==1.2.0
requests==2.32.5
pytest==7.4.4
nltk==3.8.1
